Revamp task queues
Tasks are now always on a queue. This makes for slightly nicer code,
For example, handle_receive can move tasks between queues without
knowing if it is called from the sender or receiver. Before, moving
the current task onto the run queue prematurely would cause the main
loop to corrupt the queue.