Add current implementation of poll.
It passes a very basic functional test; more sophisticated tests
will come soon.

Theory:
For each event, a channel descriptor can be in three states:
 - not polled: the channel descriptor's node for that event is a
   singleton
 - on the channel's poll_list: when the corresponding channel event
   fires, it will be triggered
 - on the task's poll_queue: the event has already been triggered, and
   the task will pick it up with a poll_wait call

Then, polling reduces to managing the channel descriptor states.

Presently, since only SEND and RECV are supported, events can be
considered "active" or "inactive". In the future, events may be
one-shot events which require special attention.

If an event is inactive at the time of poll_wait, then it is ignored.
This is required because active events are immediately enqueued again
after they are dequeued with poll_wait, since the event is still active
(and there is no guarantee that the task will change that). By contrast,
a one-shot event should not be enqueued after being dequeued in
poll_wait, and should be unconditionally sent to the user (unlike
"inactive" events which are ignored).